+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+### Added
+- Alpine apk repositories are now supported
+
 ## [0.3] - 2024-06-02
 
 ### Added